Switch int to long types

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[] {1, 2, 8, 11});
+        SalesManager salesManager = new SalesManager(new long[] {1, 2, 8, 11});
         System.out.println(SalesManager.max());
     }
     }
