+++ src/SalesManager.java
@@ -1,17 +1,17 @@
 public class SalesManager {
-    protected static int[] sales;
+    protected static long[] sales;
 
-    public SalesManager(int[] sales) {
+    public SalesManager(long[] sales) {
         this.sales = sales;
     }
 
     public static int max() {
-        int max = -1;
-        for (int sale : sales) {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
         }
-        return max;
+        return (int) max;
     }
 }
